% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourpicker-package.R
\docType{package}
\name{colourpicker-package}
\alias{colourpicker}
\alias{colourpicker-package}
\title{colourpicker: A Colour Picker Tool for Shiny and for Selecting Colours in Plots}
\description{
A colour picker that can be used as an input in 'Shiny' apps or Rmarkdown documents. The colour picker supports alpha opacity, custom colour palettes, and many more options. A Plot Colour Helper tool is available as an 'RStudio' Addin, which helps you pick colours to use in your plots. A more generic Colour Picker 'RStudio' Addin is also provided to let you select colours to use in your R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/colourpicker}
  \item \url{https://daattali.com/shiny/colourInput/}
  \item Report bugs at \url{https://github.com/daattali/colourpicker/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID})

Other contributors:
\itemize{
  \item David Griswold \email{novachild@gmail.com} [contributor]
}

}
\keyword{internal}
