% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run a colourpicker example}
\usage{
runExample()
}
\description{
Launch a \code{colourpicker} example Shiny app that shows how to use the
\code{colourInput} control. \cr
The example is also
\href{https://daattali.com/shiny/colourInput/}{available online}.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  runExample()
}
}
