% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourpicker-package.R
\docType{package}
\name{colourpicker-package}
\alias{colourpicker-package}
\title{colourpicker}
\description{
A colour picker tool for Shiny and for selecting colours in plots
}
\details{
A colour picker that can be used as an input in Shiny apps or Rmarkdown documents. The colour picker supports alpha opacity, custom colour palettes, and many more options. A Plot Colour Helper tool is available as an RStudio Addin, which helps you pick colours to use in your plots. A more generic Colour Picker RStudio Addin is also provided to let you select colours to use in your R code.

View the \href{https://github.com/daattali/colourpicker}{full README} for more details
and to see demos.
}
